Temporary fix for issue where votes do not display when logged in
Please see issue #2.

Votes on the main vote list are meant to be hidden if the user has
already voted on them. However this also breaks the other pages that
refer to this vote because it hides it there too. This issue will be
addressed properly as it is a temporary fix which has broken the “hide
votes I’ve already voted on” feature.